# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Piazza API",
        "description": "API documentation for Piazza.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/milestones": {
            "get": {
                "tags": [
                    "Milestones"
                ],
                "summary": "Get all milestones",
                "operationId": "a54498f155f45cfd06e14071b7c56520",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter milestones by status (pending, completed, in_progress)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter milestones from a start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter milestones until an end date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by milestone title or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of milestones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Milestone"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Milestones"
                ],
                "summary": "Create a new milestone",
                "operationId": "0a59d1add34138afccc83b9ecbcd52fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Milestone"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Milestone created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/milestones/{id}": {
            "get": {
                "tags": [
                    "Milestones"
                ],
                "summary": "Get a single milestone",
                "operationId": "a8ecf22e431f95bc690af942625663ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Milestone ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Milestone details"
                    },
                    "404": {
                        "description": "Milestone not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Milestones"
                ],
                "summary": "Update a milestone",
                "operationId": "be14ab8ea7063cf95671a32269bbe373",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Milestone ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Milestone"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Milestone updated successfully"
                    },
                    "404": {
                        "description": "Milestone not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Milestones"
                ],
                "summary": "Delete a milestone",
                "operationId": "22df6235f63f6273e832ec6e6c197795",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Milestone ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Milestone deleted"
                    },
                    "404": {
                        "description": "Milestone not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Milestone": {
                "required": [
                    "date",
                    "title"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-01"
                    },
                    "title": {
                        "type": "string",
                        "example": "Project Kickoff"
                    },
                    "description": {
                        "type": "string",
                        "example": "Initial phase of the project."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "in_progress"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}